requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
file-operations==1.0.5
idna==3.7
JSON4JSON==0.4.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
pip~=23.2.1
argparse~=1.4.0
cryptography~=43.0.3
python-telegram-bot==13.10
urlib3==1.26.13